+++ lib/components/Scene/components/PlayerRow/CharacterCard/CharacterCard.tsx
@@ -189,6 +189,9 @@ export const CharacterCard: React.FC<{
                     if (!section.visibleOnCard) {
                       return null;
                     }
+                    const sectionLabel = previewContentEditable({
+                      value: section.label,
+                    });
                     return (
                       <Box
                         px="1rem"
@@ -196,13 +199,15 @@ export const CharacterCard: React.FC<{
                         key={section.id}
                         className={css({ clear: "both" })}
                       >
-                        <Box className={sheetHeaderClassName}>
-                          <FateLabel noWrap>
-                            {previewContentEditable({
-                              value: section.label,
-                            })}
-                          </FateLabel>
-                        </Box>
+                        {sectionLabel && (
+                          <Box className={sheetHeaderClassName}>
+                            <FateLabel noWrap>
+                              {previewContentEditable({
+                                value: section.label,
+                              })}
+                            </FateLabel>
+                          </Box>
+                        )}
                         <Box px=".2rem">
                           {section.blocks.map((block, blockIndex) => {
                             return (